requirements.txt
discord.py>=2.0.0
requests>=2.28.0
python-dotenv>=0.19.0
PyQt5>=5.15.0
Pillow>=8.0.0
pytesseract>=0.3.8
mss>=6.1.0
googletrans>=3.1.0a0
pyautogui>=0.9.53